This function test if an 'sf' object match with Uruguay at crs = 4326.

This function test if an 'sf' object match with Uruguay at crs = 4326.

is.uy4326(x)

Arguments

x
An 'sf' object with the same crs as the homonym parameter

Value

logical value based in crs parameter of the sf object

See also

Examples

# \donttest{
is.uy4326(load_geouy("Peajes"))

#> Reading layer `peajes' from data source `https://geoservicios.mtop.gub.uy/geoserver/inf_tte_ttelog_terrestre/peajes/ows?service=WFS&request=GetFeature&typeName=peajes' using driver `GML'
#> Simple feature collection with 18 features and 4 fields
#> geometry type:  MULTIPOINT
#> dimension:      XY
#> bbox:           xmin: -58.24649 ymin: -34.78446 xmax: -54.42373 ymax: -31.54664
#> geographic CRS: WGS 84
#> Reading layer `ine_pais_corregido' from data source `https://mapas.mides.gub.uy:443/geoserver/INECenso2004/ows?service=WFS&version=1.0.0&request=GetFeature&typeName=INECenso2004:ine_pais_corregido' using driver `GML'
#> Simple feature collection with 1 feature and 1 field
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: 366582.2 ymin: 6127919 xmax: 858252.1 ymax: 6671738
#> projected CRS:  WGS 84 / UTM zone 21S

#> [1] "The object you want to process is not at CRS 4326"

# }