This function test if an 'sf' object match with Uruguay at crs = 5381.

This function test if an 'sf' object match with Uruguay at crs = 5381.

is.uy5381(x)

Arguments

x

An 'sf' object with the same crs as the homonym parameter

Value

logical value based in crs parameter of the sf object

See also

Examples

# \donttest{
is.uy5381(load_geouy("CCZ"))
#> Reading layer `ZonasComunales' from data source 
#>   `https://mapas.mides.gub.uy/geoserver/ows?service=WFS&version=1.0.0&request=GetFeature&typeName=IDE:ZonasComunales' 
#>   using driver `GML'
#> Simple feature collection with 18 features and 3 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 551994.3 ymin: 6133494 xmax: 589199.4 ymax: 6159799
#> Projected CRS: WGS 84 / UTM zone 21S
#> Reading layer `ine_pais_corregido' from data source 
#>   `https://mapas.mides.gub.uy:443/geoserver/INECenso2004/ows?service=WFS&version=1.0.0&request=GetFeature&typeName=INECenso2004:ine_pais_corregido' 
#>   using driver `GML'
#> Simple feature collection with 1 feature and 1 field
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 366582.2 ymin: 6127919 xmax: 858252.1 ymax: 6671738
#> Projected CRS: WGS 84 / UTM zone 21S
#> [1] "The object you want to process is not at CRS 5381"
# }