This function allows to take oficial uruguayan geometries, as object "sf", from various servers.

This function allows to take oficial uruguayan geometries, as object "sf", from various servers.

load_geouy(c, crs = 32721, folder = tempdir())

Arguments

c

Define the geometries to download: may be: "Departamentos", "Secciones", "Zonas", etc. View(metadata) for details.

crs

Define the Coordinate Reference Systems you want the output, default 32721

folder

Folder where are the files download if formato == "zip" in metadata. Default tempdir()

Value

sf object with the requested geometries

See also

Examples

# \donttest{
secc <- try(load_geouy(c = "Secciones"), silent = TRUE)
#> Reading layer `Secciones' from data source 
#>   `https://mapas.mides.gub.uy:443/geoserver/ows?service=WFS&version=1.0.0&request=GetFeature&typeName=INECenso:Secciones' 
#>   using driver `GML'
#> Simple feature collection with 236 features and 10 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 366605.4 ymin: 6121402 xmax: 858003.8 ymax: 6671734
#> Projected CRS: WGS 84 / UTM zone 21S
if (!inherits(secc, "try-error")) head(secc)
#> Simple feature collection with 6 features and 10 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 612797.8 ymin: 6173795 xmax: 829630.7 ymax: 6414470
#> Projected CRS: WGS 84 / UTM zone 21S
#>                                     gml_id id DEPTO SECCION CODSEC   NOMBDEPTO
#> 1 Secciones.fid--7a08045e_1a08845f831_776d  1    10       8   1008   MALDONADO
#> 2 Secciones.fid--7a08045e_1a08845f831_776e  2     6       7    607     DURAZNO
#> 3 Secciones.fid--7a08045e_1a08845f831_776f  3     4       2    402 CERRO LARGO
#> 4 Secciones.fid--7a08045e_1a08845f831_7770  4     6       8    608     DURAZNO
#> 5 Secciones.fid--7a08045e_1a08845f831_7771  5    10       9   1009   MALDONADO
#> 6 Secciones.fid--7a08045e_1a08845f831_7772  6     6       9    609     DURAZNO
#>   viv_tot_23 pob_tot_23 pob_tot_hom_23 pob_tot_muj_23
#> 1       2189       2942           1412           1530
#> 2        973       1441            761            680
#> 3        774        932            513            419
#> 4        766       1341            656            685
#> 5        451        191            115             76
#> 6       3809       7495           3648           3847
#>                         the_geom
#> 1 MULTIPOLYGON (((705073.2 62...
#> 2 MULTIPOLYGON (((632031.5 63...
#> 3 MULTIPOLYGON (((807381.4 63...
#> 4 MULTIPOLYGON (((654689.5 63...
#> 5 MULTIPOLYGON (((715826.9 61...
#> 6 MULTIPOLYGON (((624373.8 63...
# }