This function allows to add to an 'sf' object its spatial coincidence with one or more administrative units in Uruguay, generating the corresponding variables.

This function allows to add to an 'sf' object its spatial coincidence with one or more administrative units in Uruguay, generating the corresponding variables.

which_uy(x, c = c("Localidades pg", "Departamentos"), d = c("cod", "name"))

Arguments

x
An 'sf' object with the same crs as the homonym parameter

c
Define the geometries to download: may be: "Departamentos", "Secciones", "Zonas", etc. View(metadata) for details.

d
A vector who determines the variables to be added, with three options: "cod", "name", or "full". Default c("cod", "name").

Value

sf object with the x geometries, with d variables requested from c added

Examples

# \donttest{
x <- load_geouy("Peajes")

#> Reading layer `peajes' from data source `https://geoservicios.mtop.gub.uy/geoserver/inf_tte_ttelog_terrestre/peajes/ows?service=WFS&request=GetFeature&typeName=peajes' using driver `GML'
#> Simple feature collection with 18 features and 4 fields
#> geometry type:  MULTIPOINT
#> dimension:      XY
#> bbox:           xmin: -58.24649 ymin: -34.78446 xmax: -54.42373 ymax: -31.54664
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs

x1 <- which_uy(x, c = "Localidades pg")

#> Reading layer `Localidades_pg' from data source `https://mapas.mides.gub.uy:443/geoserver/ows?service=WFS&version=1.0.0&request=GetFeature&typeName=INECenso:Localidades_pg' using driver `GML'
#> Simple feature collection with 615 features and 10 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: 367295 ymin: 6127919 xmax: 850978.1 ymax: 6661441
#> epsg (SRID):    32721
#> proj4string:    +proj=utm +zone=21 +south +datum=WGS84 +units=m +no_defs

# }